test: extend existing vendor invoice preset suites
