+++ packages/backend/test/vendorInvoiceEditPolicyEnforcementPreset.test.js
@@ -81,12 +81,12 @@ function vendorInvoiceBase(overrides = {}) {
   };
 }
 
-function withVendorInvoicePolicyEnv(fn) {
+function withVendorInvoicePolicyEnv(preset, fn) {
   return withEnv(
     {
       DATABASE_URL: process.env.DATABASE_URL || MIN_DATABASE_URL,
       AUTH_MODE: 'header',
-      ACTION_POLICY_ENFORCEMENT_PRESET: 'phase2_core',
+      ACTION_POLICY_ENFORCEMENT_PRESET: preset,
       ACTION_POLICY_REQUIRED_ACTIONS: '',
     },
     fn,
@@ -109,250 +109,253 @@ function allowPolicy(actionKey) {
   ];
 }
 
-test('PUT /vendor-invoices/:id/allocations: phase2_core required action denies when policy is missing', async () => {
-  await withVendorInvoicePolicyEnv(async () => {
-    let transactionCalled = 0;
-    await withPrismaStubs(
-      {
-        'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
-        'actionPolicy.findMany': async () => [],
-        $transaction: async () => {
-          transactionCalled += 1;
-          throw new Error('unexpected transaction in deny path');
+for (const preset of ['phase2_core', 'phase3_strict']) {
+  test(`PUT /vendor-invoices/:id/allocations: ${preset} required action denies when policy is missing`, async () => {
+    await withVendorInvoicePolicyEnv(preset, async () => {
+      let transactionCalled = 0;
+      await withPrismaStubs(
+        {
+          'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
+          'actionPolicy.findMany': async () => [],
+          $transaction: async () => {
+            transactionCalled += 1;
+            throw new Error('unexpected transaction in deny path');
+          },
         },
-      },
-      async () => {
-        const server = await buildServer({ logger: false });
-        try {
-          const res = await server.inject({
-            method: 'PUT',
-            url: '/vendor-invoices/vi-001/allocations',
-            headers: adminHeaders(),
-            payload: {
-              allocations: [
-                {
-                  projectId: 'proj-001',
-                  amount: 1000,
-                  taxRate: 0.1,
-                  taxAmount: 100,
-                },
-              ],
-            },
-          });
-          assert.equal(res.statusCode, 403, res.body);
-          const payload = JSON.parse(res.body);
-          assert.equal(payload?.error?.code, 'ACTION_POLICY_DENIED');
-          assert.equal(transactionCalled, 0);
-        } finally {
-          await server.close();
-        }
-      },
-    );
+        async () => {
+          const server = await buildServer({ logger: false });
+          try {
+            const res = await server.inject({
+              method: 'PUT',
+              url: '/vendor-invoices/vi-001/allocations',
+              headers: adminHeaders(),
+              payload: {
+                allocations: [
+                  {
+                    projectId: 'proj-001',
+                    amount: 1000,
+                    taxRate: 0.1,
+                    taxAmount: 100,
+                  },
+                ],
+              },
+            });
+            assert.equal(res.statusCode, 403, res.body);
+            const payload = JSON.parse(res.body);
+            assert.equal(payload?.error?.code, 'ACTION_POLICY_DENIED');
+            assert.equal(transactionCalled, 0);
+          } finally {
+            await server.close();
+          }
+        },
+      );
+    });
   });
-});
 
-test('PUT /vendor-invoices/:id/allocations: policy allow reaches downstream update path (not ACTION_POLICY_DENIED)', async () => {
-  await withVendorInvoicePolicyEnv(async () => {
-    let transactionCalled = 0;
-    let updateCalled = 0;
-    let deleteManyCalled = 0;
-    let createManyCalled = 0;
-    let currentAllocations = [];
+  test(`PUT /vendor-invoices/:id/allocations: ${preset} policy allow reaches downstream update path`, async () => {
+    await withVendorInvoicePolicyEnv(preset, async () => {
+      let transactionCalled = 0;
+      let updateCalled = 0;
+      let deleteManyCalled = 0;
+      let createManyCalled = 0;
+      let currentAllocations = [];
 
-    const tx = {
-      vendorInvoiceAllocation: {
-        deleteMany: async () => {
-          deleteManyCalled += 1;
-          currentAllocations = [];
-          return { count: 1 };
+      const tx = {
+        vendorInvoiceAllocation: {
+          deleteMany: async () => {
+            deleteManyCalled += 1;
+            currentAllocations = [];
+            return { count: 1 };
+          },
+          createMany: async ({ data }) => {
+            createManyCalled += 1;
+            currentAllocations = data.map((entry, index) => ({
+              id: `alloc-${index + 1}`,
+              createdAt: new Date(`2026-01-0${index + 1}T00:00:00.000Z`),
+              ...entry,
+            }));
+            return { count: data.length };
+          },
         },
-        createMany: async ({ data }) => {
-          createManyCalled += 1;
-          currentAllocations = data.map((entry, index) => ({
-            id: `alloc-${index + 1}`,
-            createdAt: new Date(`2026-01-0${index + 1}T00:00:00.000Z`),
-            ...entry,
-          }));
-          return { count: data.length };
+        vendorInvoice: {
+          update: async ({ where }) => {
+            updateCalled += 1;
+            return { id: where.id };
+          },
         },
-      },
-      vendorInvoice: {
-        update: async ({ where }) => {
-          updateCalled += 1;
-          return { id: where.id };
-        },
-      },
-    };
+      };
 
-    await withPrismaStubs(
-      {
-        'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
-        'actionPolicy.findMany': async () => allowPolicy('update_allocations'),
-        'project.findMany': async () => [{ id: 'proj-001' }],
-        'vendorInvoiceAllocation.findMany': async () => currentAllocations,
-        $transaction: async (callback) => {
-          transactionCalled += 1;
-          return callback(tx);
+      await withPrismaStubs(
+        {
+          'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
+          'actionPolicy.findMany': async () =>
+            allowPolicy('update_allocations'),
+          'project.findMany': async () => [{ id: 'proj-001' }],
+          'vendorInvoiceAllocation.findMany': async () => currentAllocations,
+          $transaction: async (callback) => {
+            transactionCalled += 1;
+            return callback(tx);
+          },
+          'auditLog.create': async () => ({ id: 'audit-alloc-001' }),
         },
-        'auditLog.create': async () => ({ id: 'audit-alloc-001' }),
-      },
-      async () => {
-        const server = await buildServer({ logger: false });
-        try {
-          const res = await server.inject({
-            method: 'PUT',
-            url: '/vendor-invoices/vi-001/allocations',
-            headers: adminHeaders(),
-            payload: {
-              allocations: [
-                {
-                  projectId: 'proj-001',
-                  amount: 1000,
-                  taxRate: 0.1,
-                  taxAmount: 100,
-                },
-              ],
-            },
-          });
-          assert.equal(res.statusCode, 200, res.body);
-          const payload = JSON.parse(res.body);
-          assert.equal(payload?.items?.length, 1);
-          assert.equal(payload?.totals?.grossTotal, 1100);
-          assert.equal(transactionCalled, 1);
-          assert.equal(deleteManyCalled, 1);
-          assert.equal(createManyCalled, 1);
-          assert.equal(updateCalled, 1);
-        } finally {
-          await server.close();
-        }
-      },
-    );
+        async () => {
+          const server = await buildServer({ logger: false });
+          try {
+            const res = await server.inject({
+              method: 'PUT',
+              url: '/vendor-invoices/vi-001/allocations',
+              headers: adminHeaders(),
+              payload: {
+                allocations: [
+                  {
+                    projectId: 'proj-001',
+                    amount: 1000,
+                    taxRate: 0.1,
+                    taxAmount: 100,
+                  },
+                ],
+              },
+            });
+            assert.equal(res.statusCode, 200, res.body);
+            const payload = JSON.parse(res.body);
+            assert.equal(payload?.items?.length, 1);
+            assert.equal(payload?.totals?.grossTotal, 1100);
+            assert.equal(transactionCalled, 1);
+            assert.equal(deleteManyCalled, 1);
+            assert.equal(createManyCalled, 1);
+            assert.equal(updateCalled, 1);
+          } finally {
+            await server.close();
+          }
+        },
+      );
+    });
   });
-});
 
-test('PUT /vendor-invoices/:id/lines: phase2_core required action denies when policy is missing', async () => {
-  await withVendorInvoicePolicyEnv(async () => {
-    let transactionCalled = 0;
-    await withPrismaStubs(
-      {
-        'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
-        'actionPolicy.findMany': async () => [],
-        $transaction: async () => {
-          transactionCalled += 1;
-          throw new Error('unexpected transaction in deny path');
+  test(`PUT /vendor-invoices/:id/lines: ${preset} required action denies when policy is missing`, async () => {
+    await withVendorInvoicePolicyEnv(preset, async () => {
+      let transactionCalled = 0;
+      await withPrismaStubs(
+        {
+          'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
+          'actionPolicy.findMany': async () => [],
+          $transaction: async () => {
+            transactionCalled += 1;
+            throw new Error('unexpected transaction in deny path');
+          },
         },
-      },
-      async () => {
-        const server = await buildServer({ logger: false });
-        try {
-          const res = await server.inject({
-            method: 'PUT',
-            url: '/vendor-invoices/vi-001/lines',
-            headers: adminHeaders(),
-            payload: {
-              lines: [
-                {
-                  lineNo: 1,
-                  description: 'Service A',
-                  quantity: 1,
-                  unitPrice: 1000,
-                  amount: 1000,
-                  taxRate: 0.1,
-                  taxAmount: 100,
-                },
-              ],
-            },
-          });
-          assert.equal(res.statusCode, 403, res.body);
-          const payload = JSON.parse(res.body);
-          assert.equal(payload?.error?.code, 'ACTION_POLICY_DENIED');
-          assert.equal(transactionCalled, 0);
-        } finally {
-          await server.close();
-        }
-      },
-    );
+        async () => {
+          const server = await buildServer({ logger: false });
+          try {
+            const res = await server.inject({
+              method: 'PUT',
+              url: '/vendor-invoices/vi-001/lines',
+              headers: adminHeaders(),
+              payload: {
+                lines: [
+                  {
+                    lineNo: 1,
+                    description: 'Service A',
+                    quantity: 1,
+                    unitPrice: 1000,
+                    amount: 1000,
+                    taxRate: 0.1,
+                    taxAmount: 100,
+                  },
+                ],
+              },
+            });
+            assert.equal(res.statusCode, 403, res.body);
+            const payload = JSON.parse(res.body);
+            assert.equal(payload?.error?.code, 'ACTION_POLICY_DENIED');
+            assert.equal(transactionCalled, 0);
+          } finally {
+            await server.close();
+          }
+        },
+      );
+    });
   });
-});
 
-test('PUT /vendor-invoices/:id/lines: policy allow reaches downstream update path (not ACTION_POLICY_DENIED)', async () => {
-  await withVendorInvoicePolicyEnv(async () => {
-    let transactionCalled = 0;
-    let updateCalled = 0;
-    let deleteManyCalled = 0;
-    let createManyCalled = 0;
-    let currentLines = [];
+  test(`PUT /vendor-invoices/:id/lines: ${preset} policy allow reaches downstream update path`, async () => {
+    await withVendorInvoicePolicyEnv(preset, async () => {
+      let transactionCalled = 0;
+      let updateCalled = 0;
+      let deleteManyCalled = 0;
+      let createManyCalled = 0;
+      let currentLines = [];
 
-    const tx = {
-      vendorInvoiceLine: {
-        deleteMany: async () => {
-          deleteManyCalled += 1;
-          currentLines = [];
-          return { count: 1 };
+      const tx = {
+        vendorInvoiceLine: {
+          deleteMany: async () => {
+            deleteManyCalled += 1;
+            currentLines = [];
+            return { count: 1 };
+          },
+          createMany: async ({ data }) => {
+            createManyCalled += 1;
+            currentLines = data.map((entry, index) => ({
+              id: `line-${index + 1}`,
+              createdAt: new Date(`2026-01-0${index + 1}T00:00:00.000Z`),
+              ...entry,
+            }));
+            return { count: data.length };
+          },
         },
-        createMany: async ({ data }) => {
-          createManyCalled += 1;
-          currentLines = data.map((entry, index) => ({
-            id: `line-${index + 1}`,
-            createdAt: new Date(`2026-01-0${index + 1}T00:00:00.000Z`),
-            ...entry,
-          }));
-          return { count: data.length };
+        vendorInvoice: {
+          update: async ({ where }) => {
+            updateCalled += 1;
+            return { id: where.id };
+          },
         },
-      },
-      vendorInvoice: {
-        update: async ({ where }) => {
-          updateCalled += 1;
-          return { id: where.id };
-        },
-      },
-    };
+      };
 
-    await withPrismaStubs(
-      {
-        'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
-        'actionPolicy.findMany': async () => allowPolicy('update_lines'),
-        'vendorInvoiceLine.groupBy': async () => [],
-        'vendorInvoiceLine.findMany': async () => currentLines,
-        $transaction: async (callback) => {
-          transactionCalled += 1;
-          return callback(tx);
+      await withPrismaStubs(
+        {
+          'vendorInvoice.findUnique': async () => vendorInvoiceBase(),
+          'actionPolicy.findMany': async () => allowPolicy('update_lines'),
+          'vendorInvoiceLine.groupBy': async () => [],
+          'vendorInvoiceLine.findMany': async () => currentLines,
+          $transaction: async (callback) => {
+            transactionCalled += 1;
+            return callback(tx);
+          },
+          'auditLog.create': async () => ({ id: 'audit-line-001' }),
         },
-        'auditLog.create': async () => ({ id: 'audit-line-001' }),
-      },
-      async () => {
-        const server = await buildServer({ logger: false });
-        try {
-          const res = await server.inject({
-            method: 'PUT',
-            url: '/vendor-invoices/vi-001/lines',
-            headers: adminHeaders(),
-            payload: {
-              lines: [
-                {
-                  lineNo: 1,
-                  description: 'Service A',
-                  quantity: 1,
-                  unitPrice: 1000,
-                  amount: 1000,
-                  taxRate: 0.1,
-                  taxAmount: 100,
-                },
-              ],
-            },
-          });
-          assert.equal(res.statusCode, 200, res.body);
-          const payload = JSON.parse(res.body);
-          assert.equal(payload?.items?.length, 1);
-          assert.equal(payload?.totals?.grossTotal, 1100);
-          assert.equal(transactionCalled, 1);
-          assert.equal(deleteManyCalled, 1);
-          assert.equal(createManyCalled, 1);
-          assert.equal(updateCalled, 1);
-        } finally {
-          await server.close();
-        }
-      },
-    );
+        async () => {
+          const server = await buildServer({ logger: false });
+          try {
+            const res = await server.inject({
+              method: 'PUT',
+              url: '/vendor-invoices/vi-001/lines',
+              headers: adminHeaders(),
+              payload: {
+                lines: [
+                  {
+                    lineNo: 1,
+                    description: 'Service A',
+                    quantity: 1,
+                    unitPrice: 1000,
+                    amount: 1000,
+                    taxRate: 0.1,
+                    taxAmount: 100,
+                  },
+                ],
+              },
+            });
+            assert.equal(res.statusCode, 200, res.body);
+            const payload = JSON.parse(res.body);
+            assert.equal(payload?.items?.length, 1);
+            assert.equal(payload?.totals?.grossTotal, 1100);
+            assert.equal(transactionCalled, 1);
+            assert.equal(deleteManyCalled, 1);
+            assert.equal(createManyCalled, 1);
+            assert.equal(updateCalled, 1);
+          } finally {
+            await server.close();
+          }
+        },
+      );
+    });
   });
-});
+}
